src/requirements.txt
aiofiles==0.5.0
aiohttp==3.6.3
alembic==1.5.4
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorlog==4.4.0
colorzero==1.1
cycler==0.10.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gpiozero==1.5.1
gunicorn==20.0.4
h11==0.11.0
idna==2.8
imutils==0.5.3
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Mako==1.1.4
MarkupSafe==1.1.1
msgpack==1.0.0
msgpack-numpy==0.4.7.1
mss==6.0.0
multidict==4.7.6
notify-run==0.0.13
numpy==1.19.5
pafy==0.5.5
phue==1.1
Pillow==8.0.0
psutil==5.8.0
pycodestyle==2.6.0
pyparsing==2.4.7
PyQRCode==1.2.1
pytesseract==0.3.6
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.22.0
rtsp==1.1.8
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.8
supervisor==4.2.1
toml==0.10.1
tqdm==4.50.2
urllib3==1.25.11
uvicorn==0.12.2
vidgear==0.1.9
visitor==0.1.3
Werkzeug==1.0.1
wget==3.2
WTForms==2.3.3
yarl==1.5.1
youtube-dl==2020.9.20
